+++ src/index.ts
@@ -1,4 +1,4 @@
-export { Builder } from './builder'
+export { Builder, builder } from './builder'
 export { ASTERISK, DISTINCT, ALL, DEFAULT } from './singletoneConstants'
 export { LogicalOperator } from './operators'
 export { e, o, f, $ } from './functions'