+++ src/steps/RootStep.ts
@@ -1,12 +1,15 @@
-import { ALL, DISTINCT } from '../singletoneConstants'
+import { ItemInfo } from 'Non-Exported/ItemInfo'
+import { DeleteFromStep } from 'Non-Exported/steps/delete-path/DeleteFromStep'
+import { IntoStep } from 'Non-Exported/steps/IntoStep'
+import { SelectFromStep } from 'Non-Exported/steps/select-path/SelectFromStep'
+import { ALL, ASTERISK, All, DISTINCT, Distinct } from '../singletoneConstants'
 import { BuilderData } from '../builder'
-import { Table } from '../database'
+import { AliasedTable, Column, Table } from '../database'
 import { PrimitiveType } from '../models'
-import { SelectItemInfo } from '../SelectItemInfo'
 import { BaseStep } from './BaseStep'
 import { DeleteStep } from './delete-path/DeleteStep'
 import { InsertStep } from './InsertStep'
-import { SelectItem, Step } from './Step'
+import { FromItems, SelectItem, Step } from './Step'
 import { UpdateStep } from './stepInterfaces'
 import { SelectStep } from './select-path/SelectStep'
 
@@ -20,25 +23,40 @@ export class RootStep extends BaseStep {
 		return ''
 	}
 
-	select(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
+	public select(distinct: Distinct|All, ...items: (ItemInfo|SelectItem|PrimitiveType)[]): SelectStep
+	public select(...items: (ItemInfo|SelectItem|PrimitiveType)[]): SelectStep
+	public select(...items: (Distinct|All|ItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
 		return new SelectStep(this.data, this, items)
 	}
 
-	selectDistinct(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
-		return new SelectStep(this.data, this, items, DISTINCT)
+	selectDistinct(...items: (ItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
+		return new SelectStep(this.data, this, [DISTINCT, ...items])
 	}
 
-	selectAll(...items: (SelectItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
-		return new SelectStep(this.data, this, items, ALL)
+	selectAll(...items: (ItemInfo|SelectItem|PrimitiveType)[]): SelectStep {
+		return new SelectStep(this.data, this, [ALL, ...items])
+	}
+
+	selectAsteriskFrom(...tables: FromItems): SelectFromStep {
+		return new SelectStep(this.data, this, [ASTERISK]).from(...tables)
 	}
 
 	delete(): DeleteStep {
 		return new DeleteStep(this.data, this)
 	}
 
+	deleteFrom(table: Table|AliasedTable): DeleteFromStep {
+		return new DeleteStep(this.data, this).from(table)
+	}
+
 	insert(): InsertStep {
 		// TODO: code for the new way
 		return new Step(this.data, this).insert()
+	}
+
+	insertInto(table: Table, ...columns: Column[]): IntoStep {
+		// TODO: code for the new way
+		return new Step(this.data, this).insert().into(table, ...columns)
 	}
 
 	update(table: Table): UpdateStep {