+++ src/steps/select-path/SelectStep.ts
@@ -1,3 +1,4 @@
+import { MoreThanOneDistinctOrAllError } from '../../errors'
 import { AggregateFunction } from '../../AggregateFunction'
 import { Binder } from '../../binder'
 import { BuilderData } from '../../builder'
@@ -6,7 +7,7 @@ import { ItemInfo } from '../../ItemInfo'
 import { Expression, PrimitiveType } from '../../models'
 import { ReturningItem } from '../../ReturningItemInfo'
 import { SelectItemInfo } from '../../SelectItemInfo'
-import { All, Asterisk, Distinct } from '../../singletoneConstants'
+import { ALL, All, Asterisk, DISTINCT, Distinct } from '../../singletoneConstants'
 import { BaseStep } from '../BaseStep'
 import { SelectFromStep } from './SelectFromStep'
 import { FromItems, SelectItem, Step } from '../Step'
@@ -14,13 +15,33 @@ import { ReturningStep } from '../ReturningStep'
 import { TableAsterisk } from '../../TableAsterisk'
 
 export class SelectStep extends BaseStep {
+	private readonly items: (ItemInfo|SelectItem|PrimitiveType)[]
+	private readonly distinct?: Distinct|All
+
 	constructor(
 		data: BuilderData,
 		prevStep: BaseStep,
-		private readonly items: (SelectItemInfo|SelectItem|PrimitiveType)[],
-		private readonly distinct?: Distinct|All,
+		items: (Distinct|All|ItemInfo|SelectItem|PrimitiveType)[],
 	) {
 		super(data, prevStep)
+		if (items[0] instanceof Distinct) {
+			if (items.length <= 1) throw new Error('Select step must have at least one parameter after DISTINCT')
+			items.shift() //remove first item the DISTINCT item
+			// TODO: use this as type guard
+			SelectStep.throwIfMoreThanOneDistinctOrAll(items)
+			this.items = items as (ItemInfo|SelectItem|PrimitiveType)[]
+			this.distinct = DISTINCT
+		}
+
+		if (items[0] instanceof All) {
+			items.shift() //remove first item the ALL item
+			SelectStep.throwIfMoreThanOneDistinctOrAll(items)
+			this.items = items as (ItemInfo|SelectItem|PrimitiveType)[]
+			this.distinct = ALL
+		}
+
+		SelectStep.throwIfMoreThanOneDistinctOrAll(items)
+		this.items = items as (ItemInfo|SelectItem|PrimitiveType)[]
 	}
 
 	from(...tables: FromItems): SelectFromStep {
@@ -33,10 +54,16 @@ export class SelectStep extends BaseStep {
 	}
 
 	getStepStatement(): string {
-		const selectItemInfos: SelectItemInfo[] = this.items.map(it => {
-			if (it instanceof SelectItemInfo) {
+		const selectItemInfos: ItemInfo[] = this.items.map(it => {
+			if (it instanceof SelectItemInfo || it instanceof ItemInfo) {
 				return it
-			} else if (it instanceof Expression || it instanceof Column || it instanceof AggregateFunction || it instanceof Asterisk || it instanceof TableAsterisk) {
+			} else if (
+				it instanceof Expression
+				|| it instanceof Column
+				|| it instanceof AggregateFunction
+				|| it instanceof Asterisk
+				|| it instanceof TableAsterisk
+			) {
 				return new SelectItemInfo(it, undefined)
 			} else if (it instanceof Binder) {
 				if (it.no === undefined) {
@@ -62,6 +89,15 @@ export class SelectStep extends BaseStep {
 		}
 
 		return result
+	}
+
+	private static throwIfMoreThanOneDistinctOrAll(items: (Distinct|All|ItemInfo|SelectItem|PrimitiveType)[]):
+		items is (SelectItemInfo|SelectItem|PrimitiveType)[] {
+		items.forEach(it => {
+			if (it instanceof Distinct || it instanceof All)
+				throw new MoreThanOneDistinctOrAllError('You can not have more than one DISTINCT or ALL')
+		})
+		return true
 	}
 }