+++ test/exported/bitwiseOperator.test.ts
@@ -1,4 +1,4 @@
-import { Builder } from 'src'
+import { builder } from 'src'
 import { database } from 'test/database'
 
 //Alias
@@ -6,7 +6,7 @@ const table = database.s.public.t.table1
 const col4 = database.s.public.t.table1.c.col4
 
 describe('Bitwise Operators', () => {
-	const sql = new Builder(database)
+	const sql = builder(database)
 	afterEach(() => { sql.cleanUp() })
 	describe('bitwise AND', () => {
 		it('Produces [SELECT * FROM "table1" WHERE ("col4" & 1) = 0;]', () => {