+++ test/exported/column.test.ts
@@ -1,10 +1,10 @@
-import { Builder } from 'src'
+import { builder } from 'src'
 
 import { ReturningItemInfo } from 'Non-Exported/ReturningItemInfo'
 import { database } from 'test/database'
 
 describe('BooleanColumn', () => {
-	const sql = new Builder(database, { throwErrorIfDeleteHasNoCondition: false })
+	const sql = builder(database, { throwErrorIfDeleteHasNoCondition: false })
 	afterEach(() => { sql.cleanUp() })
 	it(`getColumns returns array of one item of itself`, () => {
 		const actual = database.s.public.t.table1.c.col7.getColumns()